add more basic tests, finish all event handlers

+++ src/Bechtle.A365.ConfigService.Projection/DomainEventHandlers/StructureDeletedHandler.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
 using Bechtle.A365.ConfigService.Common.DomainEvents;
 using Bechtle.A365.ConfigService.Projection.DataStorage;
 using Microsoft.Extensions.Logging;
@@ -21,6 +22,9 @@ namespace Bechtle.A365.ConfigService.Projection.DomainEventHandlers
         /// <inheritdoc />
         public async Task HandleDomainEvent(StructureDeleted domainEvent)
         {
+            if (domainEvent is null)
+                throw new ArgumentNullException(nameof(domainEvent), $"{nameof(domainEvent)} must not be null");
+
             _logger.LogInformation($"deleting structure {domainEvent.Identifier}");
 
             await _database.DeleteStructure(domainEvent.Identifier);
